+++ NEWS.md
@@ -1,6 +1,8 @@
 # elasticsearchr 0.3.0
 
 * Added support for source filtering with the `select_fields` function.
+* Added the `%info%` operator to be used for retreiving cluster and index information.
+* Added the `list_index` function for retrieving a list of all available indices.
 
 
 # elasticsearchr 0.2.3